Add generic app bar for utils folder Use generic app bar over popular monuments view mobile
