+++ lib/presentation/popular_monuments/mobile/popular_monuments_view_mobile.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:monumento/application/popular_monuments/monument_3d_model/monument_3d_model_bloc.dart';
-import 'package:monumento/presentation/popular_monuments/mobile/widgets/popular_monument_view_mobile_app_bar.dart';
+import 'package:monumento/gen/assets.gen.dart';
+import 'package:monumento/presentation/notification/desktop/notification_view_desktop.dart';
 import 'package:monumento/presentation/popular_monuments/mobile/widgets/populat_monuments_view_body_mobile.dart';
 import 'package:monumento/presentation/popular_monuments/mobile/widgets/scan_monuments_screen.dart';
 import 'package:monumento/service_locator.dart';
 import 'package:monumento/utils/app_colors.dart';
 import 'package:monumento/utils/app_text_styles.dart';
+import 'package:monumento/utils/custom_mobile_appBar.dart';
 
 class PopularMonumentsViewMobile extends StatefulWidget {
   const PopularMonumentsViewMobile({super.key});
@@ -27,7 +29,23 @@ class _PopularMonumentsViewMobileState
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: PopularMonumnetsViewMobileAppBar(),
+      appBar: CustomMobileAppBar(
+        onPressed: () async {
+          await Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (ctx) {
+                return const NotificationViewDesktop();
+              },
+            ),
+          );
+        },
+        icon: Icons.notifications_outlined,
+        logo: Assets.desktop.logoDesktop.svg(
+          width: 161,
+          height: 25,
+        ),
+      ),
       body: PopularMonumentsViewMobileBodyBlocBuilder(),
       floatingActionButton: FloatingActionButton.extended(
         onPressed: () async {
